requirements.txt
python-dotenv
pymupdf4llm
sentence-transformers
chromadb
requests
